Handle token renewal process
This changes the class around a bit to put all of the tokens under a
token scope. It uses the expire time to test if the tokens are expired
and need a renewal, it also tests that the token was renewed by looking
at the debug message (in the test case)

There was some cleanup to test.py to make it easier to test the changes
locally without having to rebuild.